Add .gitignore and files/folders necessary for roadrunner
The purpose of the .gitignore is so that local code can easily fetch from and commit to the repo without local gradle scripts getting in the way.

The trajectorysequence and util folders from roadunner have been added as they are necessary for Autonomus.java, as well as some gradle build scripts being changed as necessary for roadrunner.