Improvements to armAssembly.java and add extendToDrop to armAssemblyTeleOp.java
Remove redundant functions and cleanup code in armAssembly.java

Now, when using armAssemblyTeleOp.java:
gamepad2.dpad_down moves the lift down, retracts the arm, and closes the claw. The lift power is fixed until 1) the driver overrides with gamepad2.right_stick_y or 2) the lift has moved fully to the bottom
gamepad2_dpad_up moves the lift up, extends the arm, and closes the claw. The lift power is fixed until 1) the driver overrides with gamepad2.right_stick_y or 2) the lift has moved fully to the top (Lift.max)